Change ReplyWrite.written to 32 bits
The fuse_write_in struct only uses 32 bits to request the size of the
write, so there's no point to allocating 64 bits for the response.

+++ src/raw/reply.rs
@@ -193,13 +193,13 @@ impl From<ReplyOpen> for fuse_open_out {
 /// write reply.
 pub struct ReplyWrite {
     /// the data written.
-    pub written: u64,
+    pub written: u32,
 }
 
 impl From<ReplyWrite> for fuse_write_out {
     fn from(written: ReplyWrite) -> Self {
         fuse_write_out {
-            size: written.written as u32,
+            size: written.written,
             padding: 0,
         }
     }
